# Conflicts:
#	idepix-assembly/pom.xml
#	idepix-core/pom.xml
#	idepix-meris/pom.xml
#	idepix-meris/src/main/java/org/esa/snap/idepix/meris/IdepixMerisOp.java
#	idepix-meris/src/test/java/org/esa/snap/idepix/meris/IdepixMerisOpTest.java
#	idepix-olci/src/main/java/org/esa/snap/idepix/olci/IdepixOlciOp.java
#	idepix-olci/src/main/java/org/esa/snap/idepix/olci/IdepixOlciSlopeAspectOrientationOp.java
#	idepix-olci/src/main/javahelp/org/esa/snap/idepix/olci/docs/olci/OlciAlgorithmSpecification.html
#	idepix-olci/src/test/java/org/esa/snap/idepix/olci/IdepixOlciSlopeAspectOrientationOpTest.java
#	idepix-s2msi/pom.xml
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/S2IdepixAlgorithm.java
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/S2IdepixOp.java
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/operators/cloudshadow/PotentialCloudShadowAreaIdentifier.java
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/operators/cloudshadow/S2IdepixCloudShadowOp.java
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/operators/cloudshadow/S2IdepixPostCloudShadowOp.java
#	idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/operators/cloudshadow/S2IdepixPreCloudShadowOp.java

+++ idepix-s2msi/src/main/java/org/esa/snap/idepix/s2msi/operators/cloudshadow/MountainShadowFlagger.java
@@ -24,7 +24,7 @@ class MountainShadowFlagger {
             sunZenithIntermediate =
                     (double) sourceSunZenith * (2. * Math.pow(((90. - (double) sourceSunZenith) / 90), 3) + 1.);
         } else {
-            sunZenithIntermediate = (double) sourceSunZenith;
+            sunZenithIntermediate = sourceSunZenith;
         }
         double sunZenithRad = Math.min(89.0, sunZenithIntermediate) * MathUtils.DTOR;
         if (relativePath.length < 2) {
